Create StatfulSet instead of Deployment
This changes the type of object we create in response to the creation of
a Habitat object.

The problem with Deployments is that there is no way to automatically
assign different PersistentVolumeClaims to each Pod, which resulted in
them sharing Volumes.

StatefulSets provide persistent storage, where each Pod gets exclusive
access to a Volume by way of the PersistentVolumeClaimTemplate field in
the Spec.

Signed-off-by: Lorenzo Manacorda <lorenzo@kinvolk.io>